Moved and renamed createGCPServices
